Move error formatting to wgpu-core
